Add navigation to ttt
